WPT: Service Worker: Add unexpected state check to wait_for_state()
This patch makes wait_for_state() in test-helpers.sub.js for a service
worker WPT test check an unexpected state after adding the event
handler. Before this patch, wait_for_state() only checks the equivalence
of the state to the specific state. It might wait forever if the worker
unexpectedly skips the state or it's unexpectedly evicted.

Bug: 1000544
Change-Id: I7025fad3ba25648888a762e16ae0ded63f9c4f9f